Styling changes for smaller screens and matToolTip
