+++ client/wfprev-war/src/main/angular/src/app/components/edit-project.module.ts
@@ -1,19 +1,21 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { EditProjectComponent } from 'src/app/components/edit-project/edit-project.component';
-import { CanDeactivateGuard } from 'src/app/services/util/can-deactive.guard';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: EditProjectComponent,
-    canDeactivate: [CanDeactivateGuard]
-  }
-];
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { TextFieldModule } from '@angular/cdk/text-field';
+import { EditProjectRoutingModule } from './edit-project-routing.module';
+import { EditProjectComponent } from './edit-project/edit-project.component';
+import { ProjectDetailsComponent } from './edit-project/project-details/project-details.component';
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-  providers: [CanDeactivateGuard]
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    MatTooltipModule,
+    TextFieldModule,
+    EditProjectRoutingModule,
+    EditProjectComponent,
+    ProjectDetailsComponent 
+  ]
 })
 export class EditProjectModule {}
